add shop_name report
